pylint fixes.
